requirement.txt

numpy
opencv-python
Pillow
matplotlib
pycocotools
scipy


torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0


mmcv-full>=1.7.0

# MMDetection
mmdet>=2.28.2

openmim

tqdm